ClaimGiftCoins plus refactorings

+++ src/Lykke.Service.ReferralLinks.Core/BitCoinApi/Models/OnchainModels.cs
@@ -8,7 +8,7 @@ namespace Core.BitCoin.BitcoinApi.Models
         public Guid? TransactionId { get; set; }
         public string Address { get; set; }
         public string AssetId { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 
     public class TransferData
@@ -17,7 +17,7 @@ namespace Core.BitCoin.BitcoinApi.Models
         public string SourceAddress { get; set; }
         public string DestinationAddress { get; set; }
         public string AssetId { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 
     public class TransferAllData
@@ -32,7 +32,7 @@ namespace Core.BitCoin.BitcoinApi.Models
         public Guid? TransactionId { get; set; }
         public string Address { get; set; }
         public string AssetId { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 
     public class SwapData
@@ -40,10 +40,10 @@ namespace Core.BitCoin.BitcoinApi.Models
         public Guid? TransactionId { get; set; }
         public string Multisig1 { get; set; }
         public string AssetId1 { get; set; }
-        public double Amount1 { get; set; }
+        public decimal Amount1 { get; set; }
         public string Multisig2 { get; set; }
         public string AssetId2 { get; set; }
-        public double Amount2 { get; set; }
+        public decimal Amount2 { get; set; }
     }
 
     public class RetryData
