+++ src/Lykke.Service.ReferralLinks.Core/Domain/ReferralLink/IReferralLinksStatistics.cs
@@ -8,7 +8,7 @@ namespace Lykke.Service.ReferralLinks.Core.Domain.ReferralLink
     {
         int NumberOfInvitationsSent { get; }
         int NumberOfInvitationAccepted { get; }
-        double AmountOfCoinsDistributed { get; }
+        decimal AmountOfCoinsDistributed { get; }
         int NumberOfNewUsersBroughtIn { get; }
     }
 }
